requirements.txt
alembic==1.13.1
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
bidict==0.23.1
blinker==1.7.0
Brotli==1.1.0
build==1.1.1
CacheControl==0.14.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
coloredlogs==15.0.1
crashtest==0.4.1
cryptography==42.0.5
dill==0.3.8
distlib==0.3.8
distro==1.9.0
docutils==0.21.1
dulwich==0.21.7
exceptiongroup==1.2.0
fastapi==0.110.1
fastjsonschema==2.19.1
filelock==3.13.1
Flask==3.0.2
flatbuffers==24.3.25
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
idna==3.6
imageio==2.34.0
importlib_metadata==7.0.2
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.3.1
jeepney==0.8.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==24.3.1
lazy_loader==0.4
llvmlite==0.42.0
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
mutagen==1.47.0
networkx==3.3
ngrok==1.2.0
nh3==0.2.17
numba==0.59.1
numpy==1.26.4
onnxruntime==1.17.3
opencv-python-headless==4.9.0.80
outcome==1.3.0.post0
packaging==24.0
pexpect==4.9.0
pillow==10.3.0
pipdeptree==2.16.2
pkginfo==1.10.0
platformdirs==4.2.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.6.0
pooch==1.8.1
protobuf==5.26.1
psutil==5.9.8
ptyprocess==0.7.0
pycparser==2.21
pycryptodomex==3.20.0
pydantic==1.10.15
pyftpdlib==1.5.9
Pygments==2.17.2
PyMatting==1.1.12
pyproject_hooks==1.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-multipart==0.0.9
python-socketio==5.11.2
rapidfuzz==3.6.2
readme_renderer==43.0
redis==5.0.3
referencing==0.34.0
reflex==0.4.8
reflex-hosting-cli==0.1.12
rembg==2.0.56
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
rpds-py==0.18.0
scikit-image==0.23.1
scipy==1.13.0
SecretStorage==3.3.3
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.29
sqlmodel==0.0.16
starlette==0.37.2
starlette-admin==0.13.2
sympy==1.12
tabulate==0.9.0
tifffile==2024.2.12
tomli==2.0.1
tomlkit==0.12.4
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
trove-classifiers==2024.3.3
twine==5.0.0
typer==0.12.3
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.20.0
virtualenv==20.25.1
watchdog==4.0.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
zipp==3.17.0
